mfx_dispatchの警告対策。

+++ mfx_dispatch/src/mfx_library_iterator_linux.cpp
@@ -248,14 +248,14 @@ mfxStatus SelectImplementationType(const mfxU32 adapterNum, mfxIMPL *pImplInterf
 {
     mfx_disp_adapters* adapters = NULL;
     int adapters_num = mfx_init_adapters(&adapters);
-    if (pVendorID && pDeviceID && adapterNum < adapters_num)
+    if (pVendorID && pDeviceID && (int)adapterNum < adapters_num)
     {
         *pVendorID = adapters[adapterNum].vendor_id;
         *pDeviceID = adapters[adapterNum].device_id;
     }
     if (adapters_num) free(adapters);
 
-    if (adapterNum >= adapters_num)
+    if ((int)adapterNum >= adapters_num)
         return MFX_ERR_UNSUPPORTED;
 
     if ((*pImplInterface != MFX_IMPL_VIA_ANY) &&
@@ -383,7 +383,7 @@ mfxStatus MFXLibraryIterator::SelectDLLVersion(char *pPath, size_t pathSize,
 
 mfxIMPL MFXLibraryIterator::GetImplementationType()
 {
-    if (m_selected_adapter < 0 || m_selected_adapter >= m_adapters_num)
+    if (m_selected_adapter < 0 || m_selected_adapter >= (int)m_adapters_num)
         return MFX_ERR_UNSUPPORTED;
 
     return MFX_IMPL_VIA_VAAPI;
